refactor: move account-compatibility filter out of shared into kit
filterTokensByAccountNetworkCompatibility needs the kit-bg-owned IDBAccount
type, and shared must not depend on another OneKey package. Moved it and its
tests to packages/kit/src/utils, alongside the other unit-tested kit helpers.
No behavior change; the remaining cross-network utils in shared are back to
importing types only.
